First pass TypeScript for A-B and deps
