+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "noImplicitAny": true,
+        "sourceMap": true
+    }
+}
